techDebt(registrationRefactor): Create createEstablishment function to move all establishment creation functionality out of top level
